requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# Streamlit for web interface
streamlit>=1.31.0
streamlit-option-menu>=0.3.12

# Web requests
requests>=2.31.0

# Data processing
pandas>=2.0.0

# Azure services
azure-cosmos>=4.5.0
azure-identity>=1.15.0

# Authentication (if needed)
msal>=1.25.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# Jira integration (uncomment if needed)
# jira>=3.5.0
# atlassian-python-api>=3.41.0
